Control generation using model arguments :recycle: :zap:
Documentation on these arguments can be found here:
https://huggingface.co/docs/transformers/v4.26.0/en/main_classes/text_generation#transformers.GenerationConfig

Using eos_token_id to stop after a newline means that we avoid
generating any more than is absolutely necessary, as well as being able
to skip decoding the result between every iteration.

Using bad_words_ids means that we are always able to generate a response
avoiding the listed words, rather than not sending anything if they
happen to appear in the output.